Fix dangling uintptr bugs (#139)
* Add test calling FeatureNames at varying stack depths

Ethtool.getNames can fail when calling the ETHTOOL_GSSET_INFO ioctl
if the morestack prologue of Ethtool.ioctl grows the stack. This is
because the ssetInfo variable is stored on the stack, and uintptr is
opaque and not updated when the stack is moved.

Add a test `TestFeatureNamesStackDepth` to assert the correctness at varying
stack depths. It is expected to fail at this commit and will be fixed in
the next.

```
--- FAIL: TestFeatureNamesStackDepth (0.00s)
    ethtool_test.go:224: FeatureNames("lo") incorrect at stack depth=10: got 0 features, want 60.
```

* Fix dangling uintptr bugs

A uintptr is opaque to the runtime and so any time an unsafe.Pointer is
converted into a uintptr we must ensure that the underlying pointer
cannot be invalidated.

The documentation for
[unsafe.Pointer](https://pkg.go.dev/unsafe#Pointer) lists the safe ways
to use uintptr. The relevant subsection for this patch is 2), a uintptr has no pointer semantics and it is unsafe to treat it as a live reference except in a limited set of cases. Subsection 4) describes the exception we rely on, that converting unsafe.Pointer to uintptr in an argument is live until the call completes. We'll also store unsafe.Pointer instead of uintptr in the ifreq struct. Both types have the same underlying representation and so this is equivalent to the kernel.

+++ ethtool_cmd.go
@@ -114,7 +114,7 @@ func (e *Ethtool) CmdGet(ecmd *EthtoolCmd, intf string) (uint32, error) {
 
 	ifr := ifreq{
 		ifr_name: name,
-		ifr_data: uintptr(unsafe.Pointer(ecmd)),
+		ifr_data: unsafe.Pointer(ecmd),
 	}
 
 	_, _, ep := unix.Syscall(unix.SYS_IOCTL, uintptr(e.fd),
@@ -142,7 +142,7 @@ func (e *Ethtool) CmdSet(ecmd *EthtoolCmd, intf string) (uint32, error) {
 
 	ifr := ifreq{
 		ifr_name: name,
-		ifr_data: uintptr(unsafe.Pointer(ecmd)),
+		ifr_data: unsafe.Pointer(ecmd),
 	}
 
 	_, _, ep := unix.Syscall(unix.SYS_IOCTL, uintptr(e.fd),
@@ -171,7 +171,7 @@ func (e *Ethtool) CmdGetMapped(intf string) (map[string]uint64, error) {
 
 	ifr := ifreq{
 		ifr_name: name,
-		ifr_data: uintptr(unsafe.Pointer(&ecmd)),
+		ifr_data: unsafe.Pointer(&ecmd),
 	}
 
 	_, _, ep := unix.Syscall(unix.SYS_IOCTL, uintptr(e.fd),
